break up commands.ts
