+++ packages/salesforcedx-vscode-core/src/commands/forceAuthDevHub.ts
@@ -16,7 +16,7 @@ import {
   SfdxCommandlet,
   SfdxCommandletExecutor,
   SfdxWorkspaceChecker
-} from './commands';
+} from './util';
 
 import { getRootWorkspacePath } from '../util';
