+++ packages/salesforcedx-vscode-core/src/commands/forceDebuggerStop.ts
@@ -26,7 +26,7 @@ import {
   SfdxCommandlet,
   SfdxCommandletExecutor,
   SfdxWorkspaceChecker
-} from './commands';
+} from './util';
 
 interface QueryResponse {
   status: number;
