+++ packages/salesforcedx-vscode-core/src/commands/forceProjectCreate.ts
@@ -29,7 +29,7 @@ import {
   EmptyPreChecker,
   SfdxCommandlet,
   SfdxCommandletExecutor
-} from './commands';
+} from './util';
 
 type forceProjectCreateOptions = {
   isProjectWithManifest: boolean;
