+++ packages/salesforcedx-vscode-core/src/commands/forceSourceDelete.ts
@@ -17,7 +17,7 @@ import {
   PreconditionChecker
 } from '@salesforce/salesforcedx-utils-vscode/out/src/types';
 import * as vscode from 'vscode';
-import { SfdxCommandlet, SfdxCommandletExecutor } from './commands';
+import { SfdxCommandlet, SfdxCommandletExecutor } from './util/sfdxCommandlet';
 
 import { channelService } from '../channels';
 import { nls } from '../messages';
