+++ packages/salesforcedx-vscode-core/src/commands/forceSourceStatus.ts
@@ -12,11 +12,14 @@ import {
 import { nls } from '../messages';
 import {
   EmptyParametersGatherer,
-  FlagParameter,
   SfdxCommandlet,
   SfdxCommandletExecutor,
   SfdxWorkspaceChecker
-} from './commands';
+} from './util';
+
+interface FlagParameter<T> {
+  flag: T;
+}
 
 export enum SourceStatusFlags {
   Local = '--local',
