+++ packages/salesforcedx-vscode-core/src/commands/util/index.ts
@@ -13,4 +13,28 @@ export {
   FilePathExistsChecker,
   OverwriteComponentPrompt
 } from './postconditionCheckers';
-export { SimpleGatherer } from './parameterGatherers';
+export {
+  SimpleGatherer,
+  EmptyParametersGatherer,
+  DemoModePromptGatherer,
+  CompositeParametersGatherer,
+  FileSelection,
+  FileSelector,
+  FilePathGatherer,
+  SelectOutputDir,
+  SelectFileName,
+  SelectUsername
+} from './parameterGatherers';
+export { EmptyPostChecker } from './postconditionCheckers';
+export {
+  SfdxCommandlet,
+  SfdxCommandletExecutor,
+  FlagParameter,
+  CommandletExecutor
+} from './sfdxCommandlet';
+export {
+  SfdxWorkspaceChecker,
+  CompositePreconditionChecker,
+  DevUsernameChecker,
+  EmptyPreChecker
+} from './preconditionCheckers';
